requirements (3).txt
fastapi
uvicorn
openai
googlemaps
python-dotenv
requests
